Refactor setstate
